+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add an optional generic type map to `PubSub`. <br/>
   [@cursorsdottsx](https://github.com/cursorsdottsx) in [#245](https://github.com/apollographql/graphql-subscriptions/pull/245)
+- Replace `iterall` use with native `Symbol.asyncIterator`. <br/>
+  [@n1ru4l](https://github.com/n1ru4l) in [#232](https://github.com/apollographql/graphql-subscriptions/pull/232)
 
 ### 2.0.1 (not yet released)